# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "OCPP Central System API",
        "description": "REST API for managing OCPP 1.6 charge points and transactions",
        "contact": {
            "name": "Texnouz",
            "email": "support@texnouz.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/auth/change-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Change password endpoint",
                "operationId": "change_password",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePasswordRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Password changed"
                    },
                    "401": {
                        "description": "Invalid current password"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login endpoint",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get current user info",
                "operationId": "get_current_user",
                "responses": {
                    "200": {
                        "description": "Current user info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register endpoint (admin only or open registration)",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "User created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "409": {
                        "description": "Username or email already exists"
                    }
                }
            }
        },
...
        "/api/v1/id-tags": {
            "get": {
                "tags": [
                    "IdTags"
                ],
                "summary": "List all IdTags",
                "operationId": "list_id_tags",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filter by active status",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "nullable": true
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filter by user ID",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (1-based)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 0
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of IdTags",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "IdTags"
                ],
                "summary": "Create a new IdTag",
                "operationId": "create_id_tag",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateIdTagRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "IdTag created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "409": {
                        "description": "IdTag already exists"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
        "/api/v1/id-tags/{id_tag}": {
            "get": {
                "tags": [
                    "IdTags"
                ],
                "summary": "Get a specific IdTag",
                "operationId": "get_id_tag",
                "parameters": [
                    {
                        "name": "id_tag",
                        "in": "path",
                        "description": "The ID tag value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IdTag details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "IdTag not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "IdTags"
                ],
                "summary": "Update an IdTag",
                "operationId": "update_id_tag",
                "parameters": [
                    {
                        "name": "id_tag",
                        "in": "path",
                        "description": "The ID tag value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateIdTagRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "IdTag updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "IdTag not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "IdTags"
                ],
                "summary": "Delete an IdTag",
                "operationId": "delete_id_tag",
                "parameters": [
                    {
                        "name": "id_tag",
                        "in": "path",
                        "description": "The ID tag value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IdTag deleted"
                    },
                    "404": {
                        "description": "IdTag not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
        "/api/v1/id-tags/{id_tag}/block": {
            "post": {
                "tags": [
                    "IdTags"
                ],
                "summary": "Block an IdTag",
                "operationId": "block_id_tag",
                "parameters": [
                    {
                        "name": "id_tag",
                        "in": "path",
                        "description": "The ID tag value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IdTag blocked",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "IdTag not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
        "/api/v1/id-tags/{id_tag}/unblock": {
            "post": {
                "tags": [
                    "IdTags"
                ],
                "summary": "Unblock an IdTag",
                "operationId": "unblock_id_tag",
                "parameters": [
                    {
                        "name": "id_tag",
                        "in": "path",
                        "description": "The ID tag value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "IdTag unblocked",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "IdTag not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResponse": {
                "type": "object",
                "description": "Standard API response wrapper",
                "required": [
                    "success"
                ],
                "properties": {
                    "data": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/T"
                            }
                        ],
                        "nullable": true
                    },
                    "error": {
                        "type": "string",
                        "nullable": true
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
            "ChangePasswordRequest": {
                "type": "object",
                "description": "Change password request",
                "required": [
                    "current_password",
                    "new_password"
                ],
                "properties": {
                    "current_password": {
                        "type": "string",
                        "description": "Current password"
                    },
                    "new_password": {
                        "type": "string",
                        "description": "New password (min 8 characters)"
                    }
                }
            },
...
            "CreateIdTagRequest": {
                "type": "object",
                "description": "Create IdTag request",
                "required": [
                    "id_tag"
                ],
                "properties": {
                    "expiry_date": {
                        "type": "string",
                        "description": "Expiry date (ISO 8601 format)",
                        "nullable": true
                    },
                    "id_tag": {
                        "type": "string",
                        "description": "The ID tag value (RFID card number)"
                    },
                    "max_active_transactions": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Maximum active transactions allowed",
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "description": "Display name",
                        "nullable": true
                    },
                    "parent_id_tag": {
                        "type": "string",
                        "description": "Parent ID tag for group authorization",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "description": "Status: Accepted, Blocked, Expired, Invalid"
                    },
                    "user_id": {
                        "type": "string",
                        "description": "User ID to associate with this tag",
                        "nullable": true
                    }
                },
                "example": {
                    "id_tag": "RFID001234567890",
                    "name": "John's Card",
                    "status": "Accepted"
                }
            },
...
            "LoginRequest": {
                "type": "object",
                "description": "Login request",
                "required": [
                    "username",
                    "password"
                ],
                "properties": {
                    "password": {
                        "type": "string",
                        "description": "Password"
                    },
                    "username": {
                        "type": "string",
                        "description": "Username or email"
                    }
                },
                "example": {
                    "password": "secret123",
                    "username": "admin"
                }
            },
...
            "PaginatedResponse": {
                "type": "object",
                "description": "Paginated response",
                "required": [
                    "items",
                    "total",
                    "page",
                    "limit",
                    "total_pages"
                ],
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/T"
                        }
                    },
                    "limit": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    },
                    "total": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "total_pages": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    }
                }
            },
...
            "RegisterRequest": {
                "type": "object",
                "description": "Register request",
                "required": [
                    "username",
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "description": "Email address"
                    },
                    "password": {
                        "type": "string",
                        "description": "Password (min 8 characters)"
                    },
                    "username": {
                        "type": "string",
                        "description": "Username (3-50 characters)"
                    }
                },
                "example": {
                    "email": "user@example.com",
                    "password": "secure_password_123",
                    "username": "newuser"
                }
            },
...
            "UpdateIdTagRequest": {
                "type": "object",
                "description": "Update IdTag request",
                "properties": {
                    "expiry_date": {
                        "type": "string",
                        "description": "Expiry date (ISO 8601 format)",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "description": "Whether the tag is active",
                        "nullable": true
                    },
                    "max_active_transactions": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Maximum active transactions",
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "description": "Display name",
                        "nullable": true
                    },
                    "parent_id_tag": {
                        "type": "string",
                        "description": "Parent ID tag",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "description": "Status: Accepted, Blocked, Expired, Invalid",
                        "nullable": true
                    },
                    "user_id": {
                        "type": "string",
                        "description": "User ID",
                        "nullable": true
                    }
                }
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-Key"
            },
            "bearer_auth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Bearer token"
            }
        }
    },
...
}